refactoring

+++ src/main/java/org/example/calculator/process/IMathEq.java
@@ -1,5 +1,5 @@
 package org.example.calculator.process;
 
 public interface IMathEq {
-    double calculate(double a, double b, String operation);
+    double calculate(long a, long b, String operation);
 }
